pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic_newsletter"
version = "0.1.0"
authors = [
    {name = "Jack Teitel", email = "jackteitel@example.com"},
]
description = "An agentic workflow for parsing email newsletters and creating a super newsletter"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "google-api-python-client>=2.80.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.1.0",
    "langgraph>=0.0.16",
    "SQLAlchemy>=2.0.0",
    "python-dotenv>=1.0.0",
    "openai>=1.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
]

[tool.setuptools]
packages = ["agentic_newsletter"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88
